rework when to block in async_std
before we were creating _a lot_ of blocking async tasks. that all
depended on one another. this leads to spawning a lot of threads
with async-std's scheduler. because it throws them off to try
and prevent blocking of the entire program.

this is a decent first step into working towards moving away from
that pattern. which will hopefully lead to unnecessary work. this
moves us away from spawning lots of tasks blocking on one another
to read task files. essentially we've moved from
`async_std::path::PathBuf` to `std::path::PathBuf`. We've also
been able to cut out quite a bit of cloning of strings between
the two.

we've also moved from polling log files on two tasks, down to one
task per execution.

all in all this seems to be around on par with the previous build
(although it may be about ~0.1-~0.2 seconds faster in some cases).
(I don't expect these cases to be prevelant enough to be a blanket
statement, but it's definetely not a _slow down_).