fix no log output when waiting for body
previously docker would warn you when a docker image was taking
awhile to download, this actually broke when fixing lints because
now there is no long message printed if the _body_ takes awhile
to download. It was previously only the sending of the request.

In many cases though the time we spend waiting is waiting on the
body, so actually log warnings there too.