get rid of forks (#10)
* get rid of forks

this commit removes off of our last fork `isahc`, and upgrades all
dependencies. this commit isn't ready for an RC yet, we have an
issue with color-eyre adding in new error message we don't want
(we might have to revert that dep), also a dependency of color-eyre
brings in an old version of ansi-term, leading to two copies. however,
I'll consider that "fine" for now, as we work on other things.

then finally when it comes time to cut another rc they'll either be
fixed, reverted, or some other type of system.

* fix formatting